Start filter implementation.
Needs to be designed carefully.  AsyncTask (highly) recommended.
FilterTask should take a string and perform the filtering.  Will be
tricky to beat race conditions.  Timing might need to be involved
here, or a filter queue, where we apply each filter as we go.  A filter
should take the original files list as an input, and filter based off
what it has as a present filter as a result.  Regex Matches are what
are considered a positive result. If there is a backspace pressed, we
flush the current filter and start over with the new filter.  This
should help keep things at a minimum.